book new and controller testing
