+++ app/models/book.rb
@@ -1,41 +1,2 @@
 class Book < ApplicationRecord
-
-  def index
-    @books = Book.all
-  end
-
-  def show
-    @book = Book.find(params[:id])
-  end
-
-  def new
-    @book = Book.new
-  end
-
-  def create
-    @book = Book.new(book_params)
-    if @book.save
-      redirect_to books_path
-    else
-      render "new"
-    end
-  end
-
-  def edit
-  end
-
-  def update
-    if @book.update(book_params)
-      redirect_to books_path
-    else
-      render "edit"
-    end
-  end
-
-private
-
-def book_params
-  params.require(:book).permit(:title, :author)
-end
-
 end
