+++ app/views/books/index.html.erb
@@ -1,0 +1,6 @@
+<div class="card">
+  <div class="card-body">
+    <%= @book.title %>
+    <%= @book.author %>
+  </div>
+</div>
