+++ app/views/books/new.html.erb
@@ -1,0 +1,7 @@
+<div>
+  <%= simple_form_for @book do |f| %>
+      <%= f.input :title %>
+      <%= f.input :author %>
+      <%= f.button :submit, class: 'btn btn-primary' %>
+  <% end %>
+</div>
